Zmiana temeratur na int, wyliczanie temperatur urządzeń, brakuje wzoru na zmiane temp w pomieszczeniu

+++ FuzzyLogic/Model/ChartItems/ChartItem.cs
@@ -45,7 +45,7 @@ namespace FuzzyLogic.Model
     {
         public string Text { get; set; }
         public int X { get; set; }
-        public double Y { get; set; }
+        public int Y { get; set; }
         public System.TimeSpan Time = new System.TimeSpan();
     }
 }
